Transforming PlatformEthernet into PlatformNetwork, which further develops the idea that the entirety of the API necessary to perform the networking operations required by Tiny Alpaca Server can be delivered via an API that can be implemented for embedded and host targets, as well as mocked out.
This is supported by a refactoring of HostSockets into HostNetwork
(implementing PlatformNetworkInterface) and HostSocketInfo, which emulates
a single W5500 hardware socket.

Add an optimization to WriteBufferedWrappedClientConnection::write for the
case where the data being written is at least twice the size of the write buffer.

Refactor WriteBufferedWrappedClientConnection::flush so that it can share
a core with write.

Stubs for tests are included, but don't do anything yet.