Add host_network_main.cc, a library with main() that sets up the networking support for programs running on host.
Fix HostSocketInfo to work as a better basis for implementing
PlatformNetworkInterface; it isn't yet great; I think I need
to use a separate thread for that.

Add NetEcho sketch, a start towards demonstrating the ability to use
McuNet on host, with the aim of developing, including testing and
debugging, on host before testing on the target platform.